requirements.txt
torch
torchvision
numpy
opencv-python
Pillow
onnxruntime
gradio
git+https://github.com/arsenyinfo/EnlightenGAN-inference